+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+* [2023-03-23] Added Abstractions Project
 * [2022-08-24] Add support for Android 12
 
 ### Changed